+++ tofu/physics_tools/geometry/__init__.py
@@ -1,1 +1,1 @@
-from ._pinhole_camera_estimator import main as pinhole_camera_estimator
+from ._camera1d_estimator import main as camera1d_estimator
