+++ tofu/physics_tools/geometry/_camera1d_estimator.py
